Assistance with the network name
This patch provides assistance for identifying and, in some cases,
setting the network name. It adds some debug lines so the user can
easily identify what network names are available to their tenant. If the
tenant only has access to one network, that network name will be used.

This should resolve issues such as #16.