If the found network is a forwarded port, skip
The networks returned can sometimes include a forwarded port for ssh.
When this happens to be the first network in the returned set of
networks all of the network setup will fail, as the code will use the
wrong offset into the networks array.

Check for the forwarded port and skip if necessary.